+++ R/ggcorrmat.R
@@ -67,8 +67,8 @@
 #'   (applicable only when `lab = TRUE`).
 #' @param axis.text.x.margin.t,axis.text.x.margin.r,axis.text.x.margin.b,axis.text.x.margin.l
 #'   Margins between x-axis and the variable name texts (t: top, r: right, b:
-#'   bottom, l:left) in case the names are slanted, i.e. when the tl.srt is
-#'   between `45` and `75` (Defaults: `12`, `0`, `0`, `0`, resp.).
+#'   bottom, l:left), especially useful in case the names are slanted, i.e. when the tl.srt is
+#'   between `45` and `75` (Defaults: `0`, `0`, `0`, `0`, resp.).
 #' @param insig Character used to show specialized insignificant correlation
 #'   coefficients (`"pch"` (default) or `"blank"`). If `"blank"`, the
 #'   corresponding glyphs will be removed; if "pch" is used, characters (see
@@ -83,6 +83,8 @@
 #'   colorbar (Default: `TRUE`).
 #' @param t.margin,b.margin Margins in grid units. For more details, see
 #'   `?grid::unit()`.
+#' @param messages Decides whether messages references, notes, and warnings are
+#'   to be displayed (Default: `TRUE`).
 #'
 #' @import ggplot2
 #'
@@ -149,44 +151,45 @@
 # defining the function
 ggcorrmat <-
   function(data,
-             cor.vars,
-             cor.vars.names = NULL,
-             output = "plot",
-             type = "full",
-             method = "square",
-             corr.method = "pearson",
-             exact = FALSE,
-             continuity = TRUE,
-             beta = 0.2,
-             digits = 2,
-             sig.level = 0.05,
-             hc.order = FALSE,
-             hc.method = "complete",
-             lab = TRUE,
-             colors = c("#E69F00", "white", "#009E73"),
-             outline.color = "black",
-             ggtheme = ggplot2::theme_gray,
-             ggstatsplot.theme = TRUE,
-             title = NULL,
-             subtitle = NULL,
-             caption = NULL,
-             caption.default = TRUE,
-             lab.col = "black",
-             lab.size = 5,
-             insig = "pch",
-             pch = 4,
-             pch.col = "black",
-             pch.cex = 11,
-             tl.cex = 12,
-             tl.col = "black",
-             tl.srt = 45,
-             axis.text.x.margin.t = 12,
-             axis.text.x.margin.r = 0,
-             axis.text.x.margin.b = 0,
-             axis.text.x.margin.l = 0,
-             legend.title.margin = TRUE,
-             t.margin = unit(0, "mm"),
-             b.margin = unit(3, "mm")) {
+           cor.vars,
+           cor.vars.names = NULL,
+           output = "plot",
+           type = "full",
+           method = "square",
+           corr.method = "pearson",
+           exact = FALSE,
+           continuity = TRUE,
+           beta = 0.2,
+           digits = 2,
+           sig.level = 0.05,
+           hc.order = FALSE,
+           hc.method = "complete",
+           lab = TRUE,
+           colors = c("#E69F00", "white", "#009E73"),
+           outline.color = "black",
+           ggtheme = ggplot2::theme_gray,
+           ggstatsplot.theme = TRUE,
+           title = NULL,
+           subtitle = NULL,
+           caption = NULL,
+           caption.default = TRUE,
+           lab.col = "black",
+           lab.size = 5,
+           insig = "pch",
+           pch = 4,
+           pch.col = "black",
+           pch.cex = 11,
+           tl.cex = 12,
+           tl.col = "black",
+           tl.srt = 45,
+           axis.text.x.margin.t = 0,
+           axis.text.x.margin.r = 0,
+           axis.text.x.margin.b = 0,
+           axis.text.x.margin.l = 0,
+           legend.title.margin = TRUE,
+           t.margin = unit(0, "mm"),
+           b.margin = unit(3, "mm"),
+           messages = TRUE) {
     # ========================================== dataframe ==============================================================
     #
     # creating a dataframe out of the entered variables
@@ -214,8 +217,8 @@ ggcorrmat <-
     # ========================================== statistics ==============================================================
     #
     if (corr.method == "pearson" ||
-      corr.method == "spearman" ||
-      corr.method == "kendall") {
+        corr.method == "spearman" ||
+        corr.method == "kendall") {
       # computing correlations on all included variables
       corr.mat <-
         base::round(
@@ -250,7 +253,7 @@ ggcorrmat <-
         dplyr::mutate_if(
           .tbl = .,
           .predicate = purrr::is_bare_double,
-          .funs = ~base::round(x = ., digits = digits)
+          .funs = ~ base::round(x = ., digits = digits)
         )
 
       # creating a correlation matrix of p-values
@@ -261,8 +264,8 @@ ggcorrmat <-
     # ========================================== plot ==============================================================
     if (output == "plot") {
       if (corr.method == "pearson" ||
-        corr.method == "spearman" ||
-        corr.method == "kendall") {
+          corr.method == "spearman" ||
+          corr.method == "kendall") {
         # plotting the correlalogram
         plot <- ggcorrplot::ggcorrplot(
           corr = corr.mat,
@@ -292,7 +295,7 @@ ggcorrmat <-
         #
         # if caption is not specified, use the generic version only if caption.default is TRUE
         if (is.null(caption) &&
-          pch == 4 && isTRUE(caption.default)) {
+            pch == 4 && isTRUE(caption.default)) {
           # adding text details to the plot
           plot <- plot +
             ggplot2::labs(
@@ -323,47 +326,53 @@ ggcorrmat <-
             theme_corrmat()
         }
 
-        if (tl.srt >= 45 & tl.srt <= 75) {
-          # even if ggstatsplot theme is not overlaid, still make sure there is
-          # enough distance between the axis and the label
-          plot <- plot +
-            ggplot2::theme(
-              axis.text.x = ggplot2::element_text(
-                margin = ggplot2::margin(
-                  t = axis.text.x.margin.t,
-                  r = axis.text.x.margin.r,
-                  b = axis.text.x.margin.b,
-                  l = axis.text.x.margin.l,
-                  unit = "pt"
-                )
+        # label adjustment message for the development version of ggplot2
+        if (isTRUE(messages)) {
+          if (utils::packageVersion(pkg = "ggplot2") == "2.2.1.9000") {
+            base::message(cat(
+              crayon::red("Note:"),
+              crayon::blue(
+                "If you are using development version of ggplot2, the x-axis labels may not properly.
+                Try setting `axis.text.x.margin.t` to 12 or 14; depends on the length of your variable names."
               )
+              ))
+          }
+          }
+        # even if ggstatsplot theme is not overlaid, still make sure there is
+        # enough distance between the axis and the label
+        plot <- plot +
+          ggplot2::theme(axis.text.x = ggplot2::element_text(
+            margin = ggplot2::margin(
+              t = axis.text.x.margin.t,
+              r = axis.text.x.margin.r,
+              b = axis.text.x.margin.b,
+              l = axis.text.x.margin.l,
+              unit = "pt"
             )
-        }
+          ))
 
         # creating proper spacing between the legend.title and the colorbar
         if (isTRUE(legend.title.margin)) {
-          plot <- legend_title_margin(
-            plot = plot,
-            t.margin = t.margin,
-            b.margin = b.margin
-          )
+          plot <- legend_title_margin(plot = plot,
+                                      t.margin = t.margin,
+                                      b.margin = b.margin)
         }
-      } else if (corr.method == "robust") {
-        base::message(cat(
-          crayon::red("Error:"),
-          crayon::blue(
-            "Robust correlation matrix plot is currently not supported, only correlations and p-values are."
-          )
-        ))
-      }
+        } else if (corr.method == "robust") {
+          base::message(cat(
+            crayon::red("Error:"),
+            crayon::blue(
+              "Robust correlation matrix plot is currently not supported, only correlations and p-values are."
+            )
+          ))
+        }
     }
     # ========================================== output ==============================================================
 
     # return the desired result
     if (output == "correlations") {
       if (corr.method == "pearson" ||
-        corr.method == "spearman" ||
-        corr.method == "kendall") {
+          corr.method == "spearman" ||
+          corr.method == "kendall") {
         # correlation matrix
         corr.mat %<>%
           base::as.data.frame(x = .) %>%
@@ -374,8 +383,8 @@ ggcorrmat <-
       return(corr.mat)
     } else if (output == "p-values") {
       if (corr.method == "pearson" ||
-        corr.method == "spearman" ||
-        corr.method == "kendall") {
+          corr.method == "spearman" ||
+          corr.method == "kendall") {
         # p-value matrix
         p.mat %<>%
           base::as.data.frame(x = .) %>%
@@ -386,8 +395,8 @@ ggcorrmat <-
       return(p.mat)
     } else if (output == "plot") {
       if (corr.method == "pearson" ||
-        corr.method == "spearman" ||
-        corr.method == "kendall") {
+          corr.method == "spearman" ||
+          corr.method == "kendall") {
         # correlalogram plot
         return(plot)
       }